requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
asyncpg==0.30.0
google-cloud-tasks==2.17.1
